requirements.txt
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
gunicorn==20.0.4
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
mysql-connector-python==8.0.21
networkx==2.5
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==7.2.0
protobuf==3.13.0
pydicom==2.0.0
pynetdicom==1.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
tifffile==2020.9.29
Werkzeug==1.0.1
